Problem: Migrators don't get the same connection as ragtime
In the current implementation, the `CassandraDatabase` record creates
its own connection to the cassandra cluster to do its bookkeeping work
on the `migrations` table. This forces the users migrations code to make
its own connection to the cluster. This makes it possible for the
migrations to operate on a different cluster than the one containing
the `migrations` table.

Solution:

Extend the `ragtime/DataStore` protocol to the cassandra
`SessionManange`, and create a session once at the beginning of the
migration work.